refacekit-local/ops/apps/worker/requirements.txt
redis==5.0.7
pandas==2.2.2
python-dotenv==1.0.1
